Fixed datatype of Activity.distanceInMiles to double

+++ JPAOutTheDoor/src/main/java/com/skilldistillery/otd/entities/Activity.java
@@ -43,7 +43,7 @@ public class Activity {
 	private int difficulty;
 	
 	@Column(name="distance_miles")
-	private int distanceInMiles;
+	private double distanceInMiles;
